Extract interface
 - Move MediaScannerService to MediaScannerServiceImpl
